Make MSVC happy again.
TBR=mstarzinger@chromium.org

Review URL: https://codereview.chromium.org/539973002

git-svn-id: https://v8.googlecode.com/svn/branches/bleeding_edge@23688 ce2b1a6d-e550-0410-aec6-3dcde31c8c00

+++ src/compiler/value-numbering-reducer-unittest.cc
@@ -63,14 +63,16 @@ TEST_F(ValueNumberingReducerTest, OperatorEqualityNotIdentity) {
   }
   TRACED_FORRANGE(size_t, input_count, 0, arraysize(inputs)) {
     const SimpleOperator op1(static_cast<Operator::Opcode>(input_count),
-                             Operator::kNoProperties, input_count, 1, "op");
-    Node* n1 = graph()->NewNode(&op1, input_count, inputs);
+                             Operator::kNoProperties,
+                             static_cast<int>(input_count), 1, "op");
+    Node* n1 = graph()->NewNode(&op1, static_cast<int>(input_count), inputs);
     Reduction r1 = Reduce(n1);
     EXPECT_FALSE(r1.Changed());
 
     const SimpleOperator op2(static_cast<Operator::Opcode>(input_count),
-                             Operator::kNoProperties, input_count, 1, "op");
-    Node* n2 = graph()->NewNode(&op2, input_count, inputs);
+                             Operator::kNoProperties,
+                             static_cast<int>(input_count), 1, "op");
+    Node* n2 = graph()->NewNode(&op2, static_cast<int>(input_count), inputs);
     Reduction r2 = Reduce(n2);
     EXPECT_TRUE(r2.Changed());
     EXPECT_EQ(n1, r2.replacement());
@@ -88,16 +90,17 @@ TEST_F(ValueNumberingReducerTest, SubsequentReductionsYieldTheSameNode) {
         opcode, Operator::kNoProperties, 0, 1, "Operator"));
   }
   TRACED_FORRANGE(size_t, input_count, 0, arraysize(inputs)) {
-    const SimpleOperator op1(1, Operator::kNoProperties, input_count, 1, "op1");
-    Node* n = graph()->NewNode(&op1, input_count, inputs);
+    const SimpleOperator op1(1, Operator::kNoProperties,
+                             static_cast<int>(input_count), 1, "op1");
+    Node* n = graph()->NewNode(&op1, static_cast<int>(input_count), inputs);
     Reduction r = Reduce(n);
     EXPECT_FALSE(r.Changed());
 
-    r = Reduce(graph()->NewNode(&op1, input_count, inputs));
+    r = Reduce(graph()->NewNode(&op1, static_cast<int>(input_count), inputs));
     ASSERT_TRUE(r.Changed());
     EXPECT_EQ(n, r.replacement());
 
-    r = Reduce(graph()->NewNode(&op1, input_count, inputs));
+    r = Reduce(graph()->NewNode(&op1, static_cast<int>(input_count), inputs));
     ASSERT_TRUE(r.Changed());
     EXPECT_EQ(n, r.replacement());
   }
